+++ core/src/model/BrickWall.java
@@ -1,7 +1,7 @@
 package model;
 
 public class BrickWall extends StaticGameElement {
-    public BrickWall(int x, int y) {
+    public BrickWall(float x, float y) {
         super(x, y);
     }
 }
